# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Argus Alpha API
  version: 1.0.0-alpha
  description: Contract-first Alpha API for chat-first investing idea validation.
servers:
  - url: /api/v1
paths:
...
  /api/v1/me:
    get:
      responses:
        "200":
          description: Current user profile and preferences.
        "503":
          $ref: "#/components/responses/AuthSessionVerificationUnavailable"
    patch:
      responses:
        "200":
          description: Updated user profile and preferences.
        "503":
          $ref: "#/components/responses/AuthSessionVerificationUnavailable"
  /api/v1/me/usage:
    get:
      description: Return the authenticated user's current private-alpha message and simulation allowance truth for both UTC windows.
      responses:
        "200":
          description: Backend-owned hourly and daily limit, usage, remaining capacity, exact reset times, availability, and the most restrictive window per resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageAllowanceResponse"
        "401":
          description: Missing or invalid authentication.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Private-alpha access is unavailable for this identity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Durable allowance truth is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          $ref: "#/components/responses/AuthSessionVerificationUnavailable"
...
components:
  responses:
    AuthSessionVerificationUnavailable:
      description: Bounded Supabase session verification is temporarily unavailable; retry without redirecting to login.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            type: https://api.argus.app/problems/auth-session-verification-unavailable
            title: Session Verification Unavailable
            status: 503
            detail: Argus could not verify this session. Please try again.
            code: auth_session_verification_unavailable
            request_id: req_123
  schemas:
...
    UsageWindow:
      type: object
      required:
        - limit
        - used
        - remaining
        - period_end
      properties:
        limit:
          type: integer
          minimum: 0
        used:
          type: integer
          minimum: 0
        remaining:
          type: integer
          minimum: 0
        period_end:
          type: string
          format: date-time
    UsageAllowance:
      type: object
      description: Backend-derived allowance truth for one resource, covering both active UTC windows plus availability and the most restrictive window.
      required:
        - hour
        - day
        - available_now
        - limiting_window
      properties:
        hour:
          $ref: "#/components/schemas/UsageWindow"
        day:
          $ref: "#/components/schemas/UsageWindow"
        available_now:
          type: boolean
        limiting_window:
          type: string
          enum:
            - hour
            - day
    UsageAllowances:
      type: object
      required:
        - messages
        - backtests
      properties:
        messages:
          $ref: "#/components/schemas/UsageAllowance"
        backtests:
          $ref: "#/components/schemas/UsageAllowance"
    UsageAllowanceResponse:
      type: object
      required:
        - allowances
      properties:
        allowances:
          $ref: "#/components/schemas/UsageAllowances"
...
    Error:
      type: object
      description: RFC 9457 Problem Details with code and request_id.
...